# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "apiprueba, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/colaborador": {
      "get": {
        "tags": [
          "Colaborador"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Colaborador"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ColaboradorDtoPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/colaborador/colaboradorByUsuario/{id_usuario}": {
      "get": {
        "tags": [
          "Colaborador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_usuario",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/colaborador/colaboradorByIdentificacion/{cedula_colaborador}": {
      "get": {
        "tags": [
          "Colaborador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cedula_colaborador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/colaborador/{id_colaborador}": {
      "get": {
        "tags": [
          "Colaborador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_colaborador",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Colaborador"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_colaborador",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ColaboradorDtoPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Colaborador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_colaborador",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/colaborador/exportarColaboradores": {
      "get": {
        "tags": [
          "Colaborador"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "estadosSeleccionados",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/colaborador/csv": {
      "post": {
        "tags": [
          "Colaborador"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/rol/{id_rol}": {
      "get": {
        "tags": [
          "Rol"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_rol",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ColaboradorDtoPost": {
      "type": "object",
      "properties": {
        "nombres": {
          "maxLength": 30,
          "minLength": 0,
          "type": "string"
        },
        "apellidos": {
          "maxLength": 30,
          "minLength": 0,
          "type": "string"
        },
        "estado": {
          "maxLength": 11,
          "minLength": 0,
          "type": "string"
        },
        "fecha_Creacion": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ColaboradorDtoPut": {
      "type": "object",
      "properties": {
        "id_Colaborador": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "maxLength": 11,
          "minLength": 0,
          "type": "string"
        }
      }
    },
...
  }
}